==> 1_extract_hansard_hong_kong_jsons.ipynb <==
import json
import os
from tqdm import tqdm
import pandas as pd

directory = 'hansard_hong_kong'
df = pd.DataFrame()
for filename in tqdm(os.listdir(directory), desc='Processing'):
    if filename.endswith(".json"):
        file_path = os.path.join(directory, filename)
        with open(file_path) as file:
            data = json.load(file)
        data_df = pd.json_normalize(data,
            record_path=['rows'],
            sep='_')
        data_df['file_name'] = filename
        df = pd.concat([df, data_df], ignore_index=True)
df = df.dropna(thresh=100, axis=1)
df.info()

df.to_csv('intermediate_outputs/hansard_hong_kong.csv', index=False)
